somewhat cut useEffect into smaller functions
